requirements.txt
# LSB-AI-Detection Requirements
# Core dependencies for Low Surface Brightness feature detection
# sam2 and sam3 are installed in the conda environment

# Data processing
numpy>=1.21.0
astropy>=5.0
opencv-python>=4.5.0
pycocotools>=2.0.6
Pillow>=9.0.0
scikit-image>=0.19.0

# Deep learning
torch>=2.0.0
torchvision>=0.15.0

# Configuration and utilities
PyYAML>=6.0
tqdm>=4.60.0

# Visualization
matplotlib>=3.5.0

# Optional: for Jupyter notebooks
# jupyter>=1.0.0
# ipykernel>=6.0.0
